+++ cmp/chart/ChartModel.js
@@ -6,7 +6,6 @@
  */
 import {HoistModel} from '@xh/hoist/core';
 import {action, observable, makeObservable} from '@xh/hoist/mobx';
-import {throwIf} from '@xh/hoist/utils/js';
 import {castArray, cloneDeep, merge} from 'lodash';
 
 /**
@@ -16,6 +15,9 @@ export class ChartModel extends HoistModel {
 
     @observable.ref highchartsConfig = {};
     @observable.ref series = [];
+
+    /** @member {boolean} */
+    showContextMenu;
 
     /**
      * The HighCharts instance currently being displayed. This may be used for reading
@@ -28,13 +30,14 @@ export class ChartModel extends HoistModel {
      * @param {Object} c - ChartModel configuration.
      * @param {Object} c.highchartsConfig - The initial highchartsConfig for this chart.
      * @param {(Object|Object[])} c.series - The initial data series to be displayed.
+     * @param {Boolean} [c.showContextMenu] - true to showContextMenu.  Defaults to true.  Desktop only.
      */
-    constructor({highchartsConfig, series = [], config} = {}) {
+    constructor({highchartsConfig, series = [], showContextMenu = true} = {}) {
         super();
         makeObservable(this);
-        throwIf(config, 'ChartModel "config" has been removed. Please use "highchartsConfig" instead.');
         this.highchartsConfig = highchartsConfig;
         this.series = castArray(series);
+        this.showContextMenu = showContextMenu;
     }
 
     /**